Split off from git package the stuff provided by the git builtin plugin
