+++ configs/root-compilation.tsconfig.json
@@ -140,6 +140,9 @@
       "path": "../packages/vsx-registry/compile.tsconfig.json"
     },
     {
+      "path": "../packages/git-native/compile.tsconfig.json"
+    },
+    {
       "path": "../packages/timeline/compile.tsconfig.json"
     },
     {
