+++ examples/browser/package.json
@@ -26,6 +26,7 @@
     "@theia/filesystem": "1.16.0",
     "@theia/getting-started": "1.16.0",
     "@theia/git": "1.16.0",
+    "@theia/git-native": "1.16.0",
     "@theia/keymaps": "1.16.0",
     "@theia/markers": "1.16.0",
     "@theia/messages": "1.16.0",
