+++ examples/electron/compile.tsconfig.json
@@ -126,6 +126,9 @@
       "path": "../../packages/vsx-registry/compile.tsconfig.json"
     },
     {
+      "path": "../../packages/git-native/compile.tsconfig.json"
+    },
+    {
       "path": "../../packages/timeline/compile.tsconfig.json"
     },
     {
