+++ examples/electron/package.json
@@ -28,6 +28,7 @@
     "@theia/filesystem": "1.16.0",
     "@theia/getting-started": "1.16.0",
     "@theia/git": "1.16.0",
+    "@theia/git-native": "1.16.0",
     "@theia/keymaps": "1.16.0",
     "@theia/markers": "1.16.0",
     "@theia/messages": "1.16.0",
