+++ packages/git-native/README.md
@@ -4,7 +4,7 @@
 
 <img src='https://raw.githubusercontent.com/eclipse-theia/theia/master/logo/theia.svg?sanitize=true' alt='theia-ext-logo' width='100px' />
 
-<h2>ECLIPSE THEIA - GIT EXTENSION</h2>
+<h2>THEIA - GIT SOURCE CONTROL PROVIDER EXTENSION</h2>
 
 <hr />
 
@@ -12,17 +12,11 @@
 
 ## Description
 
-The `@theia/git` extension contributes the ability to interact with Git from the user's system.
-This package does not contribute Git to the Source Control view.  Either the VS Code Git builtin plugin or the Theia git-native package must be included.
-
-The `@theia/git` extension contributes:
-- `amend section` (at the bottom of the Source Control view) 
-- `history` (if scm-extra package is included)
-- a git api giving downstream packages full access to git commands
+The `@theia/git-native` extension is a replacement for the VS Code builtin git plugin, contributing to the Source Control view.
 
 ## Additional Information
 
-- [API documentation for `@theia/git`](https://eclipse-theia.github.io/theia/docs/next/modules/git.html)
+- [API documentation for `@theia/git`](https://eclipse-theia.github.io/theia/docs/next/modules/git-native.html)
 - [Theia - GitHub](https://github.com/eclipse-theia/theia)
 - [Theia - Website](https://theia-ide.org/)
