+++ packages/git-native/src/browser/dirty-diff/dirty-diff-contribution.ts
