+++ packages/git-native/src/browser/dirty-diff/dirty-diff-module.ts
