+++ packages/git-native/src/browser/git-preferences.ts
