+++ packages/git-native/src/browser/git-uri-label-contribution.ts
@@ -17,7 +17,7 @@
 import { injectable, inject } from '@theia/core/shared/inversify';
 import { LabelProviderContribution, LabelProvider, DidChangeLabelEvent } from '@theia/core/lib/browser/label-provider';
 import URI from '@theia/core/lib/common/uri';
-import { GIT_RESOURCE_SCHEME } from './git-resource';
+import { GIT_RESOURCE_SCHEME } from '@theia/git/lib/browser/git-resource';
 
 @injectable()
 export class GitUriLabelProviderContribution implements LabelProviderContribution {
