+++ packages/git-native/src/node/git-locator/git-locator-host.ts
