+++ packages/git/src/browser/diff/git-diff-tree-model.tsx
@@ -20,15 +20,18 @@ import URI from '@theia/core/lib/common/uri';
 import { ScmTreeModel } from '@theia/scm/lib/browser/scm-tree-model';
 import { Git, GitFileStatus } from '../../common';
 import { ScmService } from '@theia/scm/lib/browser/scm-service';
-import { GitScmProvider, GitScmFileChange } from '../git-scm-provider';
+import { GitScmFileChange } from '../git-scm-file-change';
+import { LabelProvider } from '@theia/core/lib/browser';
 import { ScmResourceGroup, ScmResource } from '@theia/scm/lib/browser/scm-provider';
 import { ScmFileChange } from '@theia/scm-extra/lib/browser/scm-file-change-node';
 import { GitResourceOpener } from './git-resource-opener';
+import { ScmRepository } from '@theia/scm/lib/browser/scm-repository';
 
 @injectable()
 export class GitDiffTreeModel extends ScmTreeModel {
 
     @inject(Git) protected readonly git: Git;
+    @inject(LabelProvider) protected readonly labelProvider: LabelProvider;
     @inject(ScmService) protected readonly scmService: ScmService;
     @inject(GitResourceOpener) protected readonly resourceOpener: GitResourceOpener;
 
@@ -48,26 +51,31 @@ export class GitDiffTreeModel extends ScmTreeModel {
         this.toDisposeOnContentChange.dispose();
         const scmRepository = this.scmService.findRepository(new URI(rootUri));
         if (scmRepository && scmRepository.provider.id === 'git') {
-            const provider = scmRepository.provider as GitScmProvider;
-            this.provider = provider;
+            this.provider = scmRepository.provider;
             this.diffOptions = diffOptions;
 
-            this.refreshRepository(provider);
-            this.toDisposeOnContentChange.push(provider.onDidChange(() => {
-                this.refreshRepository(provider);
+            this.refreshRepository(scmRepository);
+            this.toDisposeOnContentChange.push(scmRepository.provider.onDidChange(() => {
+                this.refreshRepository(scmRepository);
             }));
 
         }
     }
 
-    protected async refreshRepository(provider: GitScmProvider): Promise<void> {
-        const repository = { localUri: provider.rootUri };
+    protected async refreshRepository(scmRepository: ScmRepository): Promise<void> {
+        const provider = scmRepository.provider;
+        const repository = { localUri: scmRepository.provider.rootUri };
 
         const gitFileChanges = await this.git.diff(repository, this.diffOptions);
 
-        const group: ScmResourceGroup = { id: 'changes', label: 'Files Changed', resources: [], provider, dispose: () => {} };
+        const context = {
+            rootUri: scmRepository.provider.rootUri,
+            labelProvider: this.labelProvider,
+        };
+
+        const group: ScmResourceGroup = { id: 'changes', label: 'Files Changed', resources: [], provider, dispose: () => { } };
         const resources: ScmResource[] = gitFileChanges
-            .map(change => new GitScmFileChange(change, provider, this.diffOptions.range))
+            .map(change => new GitScmFileChange(change, context, this.diffOptions.range))
             .map(change => ({
                 sourceUri: new URI(change.uri),
                 decorations: {
@@ -79,7 +87,7 @@ export class GitDiffTreeModel extends ScmTreeModel {
                 group,
             }));
         const changesGroup = { ...group, resources };
-        this._groups = [ changesGroup ];
+        this._groups = [changesGroup];
 
         this.root = this.createTree();
     }
