+++ packages/git/src/browser/git-quick-open-service.ts
@@ -16,7 +16,7 @@
 
 import { injectable, inject, optional } from '@theia/core/shared/inversify';
 import { Git, Repository, Branch, BranchType, Tag, Remote, StashEntry } from '../common';
-import { GitRepositoryProvider } from './git-repository-provider';
+import { ScmService } from '@theia/scm/lib/browser/scm-service';
 import { MessageService } from '@theia/core/lib/common/message-service';
 import { WorkspaceService } from '@theia/workspace/lib/browser/workspace-service';
 import { GitErrorHandler } from './git-error-handler';
@@ -44,7 +44,7 @@ export class GitQuickOpenService {
     @inject(LabelProvider) protected readonly labelProvider: LabelProvider;
 
     @inject(Git) protected readonly git: Git;
-    @inject(GitRepositoryProvider) protected readonly repositoryProvider: GitRepositoryProvider;
+    @inject(ScmService) protected readonly scmService: ScmService;
     @inject(QuickInputService) @optional() protected readonly quickInputService: QuickInputService;
     @inject(MessageService) protected readonly messageService: MessageService;
     @inject(WorkspaceService) protected readonly workspaceService: WorkspaceService;
@@ -479,7 +479,9 @@ export class GitQuickOpenService {
     }
 
     private getRepository(): Repository | undefined {
-        return this.repositoryProvider.selectedRepository;
+        return this.scmService.selectedRepository
+            ? { localUri: this.scmService.selectedRepository.provider.rootUri }
+            : undefined;
     }
 
     private async getRemotes(): Promise<Remote[]> {
