+++ packages/git/src/browser/history/git-commit-detail-open-handler.ts
@@ -19,10 +19,10 @@ import { WidgetOpenHandler, WidgetOpenerOptions } from '@theia/core/lib/browser'
 import URI from '@theia/core/lib/common/uri';
 import { GitCommitDetailWidgetOptions } from './git-commit-detail-widget-options';
 import { GitCommitDetailWidget } from './git-commit-detail-widget';
-import { GitScmProvider } from '../git-scm-provider';
 
 export namespace GitCommitDetailUri {
-    export const scheme = GitScmProvider.GIT_COMMIT_DETAIL;
+    export const GIT_COMMIT_DETAIL = 'git-commit-detail-widget';
+    export const scheme = GIT_COMMIT_DETAIL;
     export function toCommitSha(uri: URI): string {
         if (uri.scheme === scheme) {
             return uri.fragment;
@@ -35,7 +35,7 @@ export type GitCommitDetailOpenerOptions = WidgetOpenerOptions & GitCommitDetail
 
 @injectable()
 export class GitCommitDetailOpenHandler extends WidgetOpenHandler<GitCommitDetailWidget> {
-    readonly id = GitScmProvider.GIT_COMMIT_DETAIL;
+    readonly id = GitCommitDetailUri.GIT_COMMIT_DETAIL;
 
     canHandle(uri: URI): number {
         try {
