+++ packages/git/src/browser/history/git-history-support.ts
@@ -15,22 +15,47 @@
  ********************************************************************************/
 
 import { inject, injectable } from '@theia/core/shared/inversify';
-import { Emitter, Disposable } from '@theia/core';
-import { Git } from '../../common';
+import { Emitter, DisposableCollection } from '@theia/core';
+import { LabelProvider } from '@theia/core/lib/browser';
+import { Git, WorkingDirectoryStatus, CommitWithChanges } from '../../common';
+import { ScmExtraSupportContribution } from '@theia/scm/lib/browser/scm-service';
 import { ScmHistorySupport, HistoryWidgetOptions } from '@theia/scm-extra/lib/browser/history/scm-history-widget';
-import { ScmHistoryCommit } from '@theia/scm-extra/lib/browser/scm-file-change-node';
-import { GitScmProvider } from '../git-scm-provider';
-import { GitRepositoryTracker } from '../git-repository-tracker';
+import { ScmHistoryCommit, ScmFileChange } from '@theia/scm-extra/lib/browser/scm-file-change-node';
+import { GitCommitDetailWidgetOptions } from './git-commit-detail-widget-options';
+import { ScmProvider } from '@theia/scm/lib/browser/scm-provider';
+import { ScmHistoryProvider } from '@theia/scm-extra/lib/browser/history';
+import { GitScmFileChange, GitScmCommit, GitScmFileChangeContext } from '../git-scm-file-change';
+import { GitCommitDetailUri } from './git-commit-detail-open-handler';
+import URI from '@theia/core/lib/common/uri';
 
 @injectable()
-export class GitHistorySupport implements ScmHistorySupport {
-
-    @inject(GitScmProvider) protected readonly provider: GitScmProvider;
+export class GitHistoryContribution implements ScmExtraSupportContribution {
     @inject(Git) protected readonly git: Git;
-    @inject(GitRepositoryTracker) protected readonly repositoryTracker: GitRepositoryTracker;
+    @inject(LabelProvider) protected readonly labelProvider: LabelProvider;
+
+    readonly id = 'git';
+
+    addExtraSupport(provider: ScmProvider): void {
+        const historySupport = new GitHistorySupport(provider, this.git, this.labelProvider);
+        (provider as ScmHistoryProvider).historySupport = historySupport;
+    };
+}
+
+export class GitHistorySupport implements ScmHistorySupport, GitScmFileChangeContext {
+
+    protected toDispose = new DisposableCollection();
+    protected workingDirectoryStatus: WorkingDirectoryStatus | undefined;
+    public rootUri: string;
+    constructor(
+        public readonly provider: ScmProvider,
+        protected readonly git: Git,
+        public readonly labelProvider: LabelProvider,
+    ) {
+        this.rootUri = provider.rootUri;
+    };
 
     async getCommitHistory(options?: HistoryWidgetOptions): Promise<ScmHistoryCommit[]> {
-        const repository = this.provider.repository;
+        const repository = { localUri: this.provider.rootUri };
         const gitOptions: Git.Options.Log = {
             uri: options ? options.uri : undefined,
             maxCount: options ? options.maxCount : undefined,
@@ -39,7 +64,7 @@ export class GitHistorySupport implements ScmHistorySupport {
 
         const commits = await this.git.log(repository, gitOptions);
         if (commits.length > 0) {
-            return commits.map(commit => this.provider.createScmHistoryCommit(commit));
+            return commits.map(commit => this.createScmHistoryCommit(commit));
         } else {
             const pathIsUnderVersionControl = !options || !options.uri || await this.git.lsFiles(repository, options.uri, { errorUnmatch: true });
             if (!pathIsUnderVersionControl) {
@@ -56,10 +81,13 @@ export class GitHistorySupport implements ScmHistorySupport {
     });
     readonly onDidChangeHistory = this.onDidChangeHistoryEmitter.event;
 
-    protected onGitEventDisposable: Disposable | undefined;
-    protected onFirstListenerAdd(): void {
-        this.onGitEventDisposable = this.repositoryTracker.onGitEvent(event => {
-            const { status, oldStatus } = event || { status: undefined, oldStatus: undefined };
+    protected async onFirstListenerAdd(): Promise<void> {
+        const repository = { localUri: this.provider.rootUri };
+        this.toDispose.push(this.provider.onDidChange(async () => {
+            const status = await this.git.status(repository);
+            const oldStatus = this.workingDirectoryStatus;
+            this.workingDirectoryStatus = status;
+
             let isBranchChanged = false;
             let isHeaderChanged = false;
             if (oldStatus) {
@@ -69,13 +97,45 @@ export class GitHistorySupport implements ScmHistorySupport {
             if (isBranchChanged || isHeaderChanged || oldStatus === undefined) {
                 this.onDidChangeHistoryEmitter.fire(undefined);
             }
-        });
+        }));
     }
 
     protected onLastListenerRemove(): void {
-        if (this.onGitEventDisposable) {
-            this.onGitEventDisposable.dispose();
-            this.onGitEventDisposable = undefined;
-        }
+        this.toDispose.dispose();
     }
+
+    public createScmHistoryCommit(gitCommit: CommitWithChanges): ScmHistoryCommit {
+        const range = {
+            fromRevision: gitCommit.sha + '~1',
+            toRevision: gitCommit.sha
+        };
+
+        const rootUri = this.rootUri;
+        const scmCommit: GitScmCommit = {
+            ...GitScmCommit.createScmCommit(gitCommit),
+            commitDetailUri: this.toCommitDetailUri(gitCommit.sha),
+            gitFileChanges: gitCommit.fileChanges.map(change => new GitScmFileChange(change, this, range)),
+            get fileChanges(): ScmFileChange[] {
+                return this.gitFileChanges;
+            },
+            get commitDetailOptions(): GitCommitDetailWidgetOptions {
+                return {
+                    rootUri: rootUri,
+                    commitSha: gitCommit.sha,
+                    commitMessage: gitCommit.summary,
+                    messageBody: gitCommit.body,
+                    authorName: gitCommit.author.name,
+                    authorEmail: gitCommit.author.email,
+                    authorDate: gitCommit.author.timestamp,
+                    authorDateRelative: gitCommit.authorDateRelative,
+                };
+            }
+        };
+        return scmCommit;
+    }
+
+    protected toCommitDetailUri(commitSha: string): URI {
+        return new URI('').withScheme(GitCommitDetailUri.GIT_COMMIT_DETAIL).withFragment(commitSha);
+    }
+
 }
