+++ packages/git/src/node/dugite-git.spec.ts
@@ -40,86 +40,6 @@ describe('git', async function (): Promise<void> {
         track.cleanupSync();
     });
 
-    describe('repositories', async () => {
-
-        it('should discover only first repository', async () => {
-
-            const root = track.mkdirSync('discovery-test-1');
-            fs.mkdirSync(path.join(root, 'A'));
-            fs.mkdirSync(path.join(root, 'B'));
-            fs.mkdirSync(path.join(root, 'C'));
-            const git = await createGit();
-            await initRepository(path.join(root, 'A'));
-            await initRepository(path.join(root, 'B'));
-            await initRepository(path.join(root, 'C'));
-            const workspaceRootUri = FileUri.create(root).toString();
-            const repositories = await git.repositories(workspaceRootUri, { maxCount: 1 });
-            expect(repositories.length).to.deep.equal(1);
-
-        });
-
-        it('should discover all nested repositories', async () => {
-
-            const root = track.mkdirSync('discovery-test-2');
-            fs.mkdirSync(path.join(root, 'A'));
-            fs.mkdirSync(path.join(root, 'B'));
-            fs.mkdirSync(path.join(root, 'C'));
-            const git = await createGit();
-            await initRepository(path.join(root, 'A'));
-            await initRepository(path.join(root, 'B'));
-            await initRepository(path.join(root, 'C'));
-            const workspaceRootUri = FileUri.create(root).toString();
-            const repositories = await git.repositories(workspaceRootUri, {});
-            expect(repositories.map(r => path.basename(FileUri.fsPath(r.localUri))).sort()).to.deep.equal(['A', 'B', 'C']);
-
-        });
-
-        it('should discover all nested repositories and the root repository which is at the workspace root', async function (): Promise<void> {
-            if (isWindows) {
-                // https://github.com/eclipse-theia/theia/issues/933
-                this.skip();
-                return;
-            }
-
-            const root = track.mkdirSync('discovery-test-3');
-            fs.mkdirSync(path.join(root, 'BASE'));
-            fs.mkdirSync(path.join(root, 'BASE', 'A'));
-            fs.mkdirSync(path.join(root, 'BASE', 'B'));
-            fs.mkdirSync(path.join(root, 'BASE', 'C'));
-            const git = await createGit();
-            await initRepository(path.join(root, 'BASE'));
-            await initRepository(path.join(root, 'BASE', 'A'));
-            await initRepository(path.join(root, 'BASE', 'B'));
-            await initRepository(path.join(root, 'BASE', 'C'));
-            const workspaceRootUri = FileUri.create(path.join(root, 'BASE')).toString();
-            const repositories = await git.repositories(workspaceRootUri, {});
-            expect(repositories.map(r => path.basename(FileUri.fsPath(r.localUri))).sort()).to.deep.equal(['A', 'B', 'BASE', 'C']);
-
-        });
-
-        it('should discover all nested repositories and the container repository', async () => {
-
-            const root = track.mkdirSync('discovery-test-4');
-            fs.mkdirSync(path.join(root, 'BASE'));
-            fs.mkdirSync(path.join(root, 'BASE', 'WS_ROOT'));
-            fs.mkdirSync(path.join(root, 'BASE', 'WS_ROOT', 'A'));
-            fs.mkdirSync(path.join(root, 'BASE', 'WS_ROOT', 'B'));
-            fs.mkdirSync(path.join(root, 'BASE', 'WS_ROOT', 'C'));
-            const git = await createGit();
-            await initRepository(path.join(root, 'BASE'));
-            await initRepository(path.join(root, 'BASE', 'WS_ROOT', 'A'));
-            await initRepository(path.join(root, 'BASE', 'WS_ROOT', 'B'));
-            await initRepository(path.join(root, 'BASE', 'WS_ROOT', 'C'));
-            const workspaceRootUri = FileUri.create(path.join(root, 'BASE', 'WS_ROOT')).toString();
-            const repositories = await git.repositories(workspaceRootUri, {});
-            const repositoryNames = repositories.map(r => path.basename(FileUri.fsPath(r.localUri)));
-            expect(repositoryNames.shift()).to.equal('BASE'); // The first must be the container repository.
-            expect(repositoryNames.sort()).to.deep.equal(['A', 'B', 'C']);
-
-        });
-
-    });
-
     describe('status', async () => {
 
         it('modifying a staged file should result in two changes', async () => {
