+++ packages/scm/src/browser/scm-provider.ts
@@ -33,7 +33,7 @@ export interface ScmProvider extends Disposable {
     readonly statusBarCommands?: ScmCommand[];
     readonly onDidChangeStatusBarCommands?: Event<ScmCommand[] | undefined>;
 
-    readonly amendSupport?: ScmAmendSupport;
+    amendSupport?: ScmAmendSupport;
 }
 
 export const ScmResourceGroup = Symbol('ScmResourceGroup');
