+++ tsconfig.json
@@ -183,6 +183,12 @@
       ],
       "@theia/ovsx-client/lib/*": [
         "dev-packages/ovsx-client/src/*"
+      ],
+      "@theia/git-native/*": [
+        "packages/git-native/*"
+      ],
+      "@theia/git-native/lib/*": [
+        "packages/git-native/src/*"
       ]
     }
   }
